TechFixGuide/backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
sqlalchemy[asyncio]==2.0.36
aiomysql==0.2.0
python-dotenv==1.0.1
python-multipart==0.0.12
easyocr==1.7.1
opencv-python-headless==4.10.0.84
numpy==1.26.4
Pillow==11.0.0
psutil==6.1.0
aiofiles==24.1.0
